Remove old liststate
It's now used only for the calendar, as that will be refactored a lot soon.
