+++ lib/blocs/calendar_cubit.dart
@@ -1,10 +1,10 @@
 import 'dart:async';
 
+import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:intl/intl.dart';
 import 'package:reaxit/api/api_repository.dart';
 import 'package:reaxit/api/exceptions.dart';
-import 'package:reaxit/blocs.dart';
 import 'package:reaxit/config.dart' as config;
 import 'package:reaxit/models.dart';
 
@@ -103,7 +103,86 @@ class CalendarEvent {
       );
 }
 
-typedef CalendarState = ListState<CalendarEvent>;
+/// Generic class to be used as state for paginated lists.
+class CalendarState extends Equatable {
+  /// The results to be shown. These are outdated if `isLoading` is true.
+  final List<CalendarEvent> results;
+
+  /// A message describing why there are no results.
+  final String? message;
+
+  /// Different results are being loaded. The results are outdated.
+  final bool isLoading;
+
+  /// More of the same results are being loaded. The results are not outdated.
+  final bool isLoadingMore;
+
+  /// The last results have been loaded. There are no more pages left.
+  final bool isDone;
+
+  bool get hasException => message != null;
+
+  const CalendarState({
+    required this.results,
+    required this.message,
+    required this.isLoading,
+    required this.isLoadingMore,
+    required this.isDone,
+  });
+
+  CalendarState copyWith({
+    List<CalendarEvent>? results,
+    String? message,
+    bool? isLoading,
+    bool? isLoadingMore,
+    bool? isDone,
+  }) =>
+      CalendarState(
+        results: results ?? this.results,
+        message: message ?? this.message,
+        isLoading: isLoading ?? this.isLoading,
+        isLoadingMore: isLoadingMore ?? this.isLoadingMore,
+        isDone: isDone ?? this.isDone,
+      );
+
+  @override
+  List<Object?> get props => [
+        results,
+        message,
+        isLoading,
+        isLoadingMore,
+        isDone,
+      ];
+
+  @override
+  String toString() {
+    return 'ListState<$CalendarEvent>(isLoading: $isLoading, isLoadingMore: $isLoadingMore,'
+        ' isDone: $isDone, message: $message, ${results.length} ${CalendarEvent}s)';
+  }
+
+  const CalendarState.loading({required this.results})
+      : message = null,
+        isLoading = true,
+        isLoadingMore = false,
+        isDone = true;
+
+  const CalendarState.loadingMore({required this.results})
+      : message = null,
+        isLoading = false,
+        isLoadingMore = true,
+        isDone = true;
+
+  const CalendarState.success({required this.results, required this.isDone})
+      : message = null,
+        isLoading = false,
+        isLoadingMore = false;
+
+  const CalendarState.failure({required String this.message})
+      : results = const [],
+        isLoading = false,
+        isLoadingMore = false,
+        isDone = true;
+}
 
 class CalendarCubit extends Cubit<CalendarState> {
   static const int firstPageSize = 20;
