+++ lib/blocs/list_state.dart
@@ -1,86 +1,5 @@
 import 'package:equatable/equatable.dart';
 
-/// Generic class to be used as state for paginated lists.
-class ListState<T> extends Equatable {
-  /// The results to be shown. These are outdated if `isLoading` is true.
-  final List<T> results;
-
-  /// A message describing why there are no results.
-  final String? message;
-
-  /// Different results are being loaded. The results are outdated.
-  final bool isLoading;
-
-  /// More of the same results are being loaded. The results are not outdated.
-  final bool isLoadingMore;
-
-  /// The last results have been loaded. There are no more pages left.
-  final bool isDone;
-
-  bool get hasException => message != null;
-
-  const ListState({
-    required this.results,
-    required this.message,
-    required this.isLoading,
-    required this.isLoadingMore,
-    required this.isDone,
-  });
-
-  ListState<T> copyWith({
-    List<T>? results,
-    String? message,
-    bool? isLoading,
-    bool? isLoadingMore,
-    bool? isDone,
-  }) =>
-      ListState<T>(
-        results: results ?? this.results,
-        message: message ?? this.message,
-        isLoading: isLoading ?? this.isLoading,
-        isLoadingMore: isLoadingMore ?? this.isLoadingMore,
-        isDone: isDone ?? this.isDone,
-      );
-
-  @override
-  List<Object?> get props => [
-        results,
-        message,
-        isLoading,
-        isLoadingMore,
-        isDone,
-      ];
-
-  @override
-  String toString() {
-    return 'ListState<$T>(isLoading: $isLoading, isLoadingMore: $isLoadingMore,'
-        ' isDone: $isDone, message: $message, ${results.length} ${T}s)';
-  }
-
-  const ListState.loading({required this.results})
-      : message = null,
-        isLoading = true,
-        isLoadingMore = false,
-        isDone = true;
-
-  const ListState.loadingMore({required this.results})
-      : message = null,
-        isLoading = false,
-        isLoadingMore = true,
-        isDone = true;
-
-  const ListState.success({required this.results, required this.isDone})
-      : message = null,
-        isLoading = false,
-        isLoadingMore = false;
-
-  const ListState.failure({required String this.message})
-      : results = const [],
-        isLoading = false,
-        isLoadingMore = false,
-        isDone = true;
-}
-
 /// Generic type for states with a paginated list of results.
 ///
 /// There are a number of subtypes:
@@ -89,8 +8,8 @@ class ListState<T> extends Equatable {
 /// * [ResultsListState] - indicates that there are results.
 ///   * [DoneListState] - indicates that there are no more results.
 ///   * [LoadingMoreListState] - indicates that we are loading more results.
-abstract class XListState<T> extends Equatable {
-  const XListState();
+abstract class ListState<T> extends Equatable {
+  const ListState();
 
   /// A convenience method to get the results if they are available.
   ///
@@ -108,11 +27,11 @@ abstract class XListState<T> extends Equatable {
   List<Object?> get props => [];
 }
 
-class LoadingListState<T> extends XListState<T> {
+class LoadingListState<T> extends ListState<T> {
   const LoadingListState();
 }
 
-class ErrorListState<T> extends XListState<T> {
+class ErrorListState<T> extends ListState<T> {
   @override
   final String message;
 
@@ -122,7 +41,7 @@ class ErrorListState<T> extends XListState<T> {
   List<Object?> get props => [message];
 }
 
-class ResultsListState<T> extends XListState<T> {
+class ResultsListState<T> extends ListState<T> {
   @override
   final List<T> results;
